+++ controllers/posts/delete.php
@@ -1,22 +1,26 @@
 <?php
 
-$postModel = createBlogPostModel();
-$idPostInput = filter_input(INPUT_GET, 'id_post', FILTER_SANITIZE_NUMBER_INT);
-$currentPage = filter_input(INPUT_GET, 'current_page', FILTER_SANITIZE_NUMBER_INT);
-$post = $postModel->getPostFromId($idPostInput);
+if (isset ($_SESSION['role']) && $_SESSION['role'] === 'admin') {
+    $postModel = createBlogPostModel();
+    $idPostInput = filter_input(INPUT_GET, 'id_post', FILTER_SANITIZE_NUMBER_INT);
+    $currentPage = filter_input(INPUT_GET, 'current_page', FILTER_SANITIZE_NUMBER_INT);
+    $post = $postModel->getPostFromId($idPostInput);
 
-if ($_SERVER['REQUEST_METHOD'] === 'POST') {
-    $idPostInput = filter_input(INPUT_POST, 'id_post', FILTER_SANITIZE_NUMBER_INT);
+    if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+        $idPostInput = filter_input(INPUT_POST, 'id_post', FILTER_SANITIZE_NUMBER_INT);
 
-    $postModel->deletePost($idPostInput);
+        $postModel->deletePost($idPostInput);
 
-    header("Location: /posts/list");
+        header("Location: /posts/list");
+    }
+
+    $twig = create_twig();
+
+    echo $twig->render('/posts/delete.html.twig', [
+        'page' => "Suppression d'un Blog Post",
+        'post' => $post,
+        'current_page' => $currentPage,
+    ]);
+} else {
+    header("Location: /users/login");
 }
-
-$twig = create_twig();
-
-echo $twig->render('/posts/delete.html.twig', [
-    'page' => "Suppression d'un Blog Post",
-    'post' => $post,
-    'current_page' => $currentPage,
-]);
